Extract login subcommands to simplify WtLogin, add exchange_emp, #833
