+++ mirai-core/build.gradle.kts
@@ -121,7 +121,7 @@ kotlin {
         jvmTest {
             dependencies {
                 implementation("org.pcap4j:pcap4j-distribution:1.8.2")
-                implementation("net.mamoe:mirai-login-solver-selenium:1.0-dev-14")
+              //  implementation("net.mamoe:mirai-login-solver-selenium:1.0-dev-14")
             }
         }
     }
