+++ mirai-core/src/commonMain/kotlin/network/protocol/packet/login/wtlogin/WtLogin20.kt
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2020 Mamoe Technologies and contributors.
+ *
+ * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
+ * Use of this source code is governed by the GNU AGPLv3 license that can be found through the following link.
+ *
+ * https://github.com/mamoe/mirai/blob/master/LICENSE
+ */
+
+package net.mamoe.mirai.internal.network.protocol.packet.login.wtlogin
+
+import kotlinx.io.core.toByteArray
+import net.mamoe.mirai.internal.network.QQAndroidClient
+import net.mamoe.mirai.internal.network.guid
+import net.mamoe.mirai.internal.network.protocol.packet.*
+import net.mamoe.mirai.internal.network.protocol.packet.login.WtLogin
+import net.mamoe.mirai.utils.md5
+
+internal object WtLogin20 : WtLoginExt {
+    operator fun invoke(
+        client: QQAndroidClient,
+        t402: ByteArray
+    ): OutgoingPacket = WtLogin.Login.buildLoginOutgoingPacket(client, bodyType = 2) { sequenceId ->
+        writeSsoPacket(client, client.subAppId, WtLogin.Login.commandName, sequenceId = sequenceId) {
+            writeOicqRequestPacket(client, EncryptMethodECDH(client.ecdh), 0x0810) {
+                writeShort(20) // subCommand
+                writeShort(4) // count of TLVs, probably ignored by server?
+                t8(2052)
+                t104(client.t104)
+                t116(client.miscBitMap, client.subSigMap)
+                t401((client.device.guid + "stMNokHgxZUGhsYp".toByteArray() + t402).md5())
+            }
+        }
+    }
+}
